requirements.txt
absl-py==2.1.0
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
auto-gptq==0.2.2
backcall==0.2.0
backoff==2.2.1
bayesian-optimization==1.4.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitsandbytes==0.39.0
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.22
click==8.1.3
clickhouse-connect==0.6.2
cma==3.3.0
colorama==0.4.6
comm==0.1.3
cryptography==41.0.1
dataclasses-json==0.5.8
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dnspython==2.3.0
duckdb==0.8.1
email-validator==2.0.0.post2
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.97.0
fastjsonschema==2.17.1
filelock==3.12.0
Flask==2.3.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.32
gpt4all==0.3.6
grpcio==1.62.1
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
InstructorEmbedding==1.0.1
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
langchain==0.0.202
langchainplus-sdk==0.0.10
llama-cpp-python==0.1.48
lxml==4.9.3
lz4==4.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nevergrad==1.0.2
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pdf2image==1.16.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.2.0
posthog==3.0.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.9
pydantic_core==2.25.1
pydeck==0.8.0
Pygments==2.15.1
PyJWT==2.8.0
pymongo==4.4.1
Pympler==1.0.1
pypdf==3.11.1
pyrsistent==0.19.3
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rouge==1.0.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torch-tb-profiler==0.4.3
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
tweepy==4.14.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.2.0
urllib3==1.26.6
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==4.0.7
XlsxWriter==3.1.2
xxhash==3.2.0
yarl==1.9.2
zhipuai==2.1.5.20230904
zipp==3.16.2
zstandard==0.21.0
